requirements.txt
# AI Interview CRM Requirements
flask==3.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
werkzeug==3.0.1
google-generativeai==0.3.2
openai-whisper==20231117
PyPDF2==3.0.1
fpdf2==2.7.6
matplotlib==3.8.2
PyJWT==2.8.0
soundfile==0.12.1
scikit-learn==1.3.2
numpy==1.26.2
requests==2.31.0
flask-cors==4.0.0